requirements.txt
evernote3==1.25.12
Flask==0.12.2
iso8601utils==0.1.2
marshmallow==3.0.0b3
mongoengine==0.13.0
msgpack-python==0.4.8
querystring_parser==1.2.3
webargs==1.8.1